requirements.txt
uvicorn==0.23.2
pylint==2.17.5
Flask==2.1.1
pytest==7.4.0
python-dotenv==0.21.1